implementation d'un moteur graphique en Swing - ajout des images à utiliser en resources

+++ src/main/java/gameengine/physicsengine/Hitbox.java
@@ -5,7 +5,7 @@ import javafx.scene.shape.Shape;
 
 
 public class Hitbox extends Rectangle {
-   public Hitbox (int x, int y , double width , double height){
+   public Hitbox (double x, double y , double width , double height){
        super();
        setX(x);
        setY(y);
